// src/shared/types.d.ts
export type CommitType = 'feat' | 'fix' | 'docs' | 'style' | 'refactor' | 'perf' | 'test' | 'chore' | 'other';
export type BranchType = 'main' | 'develop' | 'feature' | 'release' | 'hotfix' | 'custom';
export interface Author {
    name: string;
    email: string;
}
export interface Commit {
    hash: string;
    shortHash: string;
    parents: string[];
    message: string;
    shortMessage: string;
    type: CommitType;
    author: Author;
    committer: Author;
    timestamp: number;
    isMerge: boolean;
    isSquash: boolean;
    branchName?: string;
}
export interface Branch {
    name: string;
    type: BranchType;
    objectName: string;
    isHead: boolean;
    isLocal: boolean;
    isRemote: boolean;
    upstream?: string;
    color: string;
    lane: number;
}
export interface Repository {
    path: string;
    name: string;
    isValidGit: boolean;
    currentBranch: string;
    headCommit: string;
    branches: Branch[];
    totalCommits: number;
}